web: move package 'templates' to 'pages/'
